// src/constants/enum.tsx
export const enum BRANDS {
  TOSHIBA = 'TOSHIBA',
  MITSUBISHI = 'MITSUBISHI',
  DAIKIN = 'DAIKIN',
  LG = 'LG',
  SHARP = 'SHARP',
  SAMSUNG = 'SAMSUNG',
  PANASONIC = 'PANASONIC'
}

export const enum STATUS {
  ON = 'ON',
  OFF = 'OFF',
  INACTIVE = 'INACTIVE'
}

export const enum MODE {
  COOLING = 'COOLING',
  DEFAULT = 'DEFAULT',
  MOISTURING = 'MOISTURING',
  NONE = 'NONE'
}